Reorganizing how the commands are ordered in the python files
