+++ gui_control.py
@@ -4,16 +4,17 @@ Script used to initiate the GUI server instance. The main documentation will be
 given in the files of the server/ directory
 """
 import server.session as ss
+
+# Loading the base command session
 import ctlcmd.cmdbase as cmdbase
-import ctlcmd.motioncmd as motioncmd
-import ctlcmd.getset as getset
+import ctlcmd.gcodercmd as gcodercmd
 import ctlcmd.digicmd as digicmd
-import ctlcmd.viscmd as viscmd
+import ctlcmd.visualcmd as visualcmd
 import ctlcmd.picocmd as picocmd
 import ctlcmd.drscmd as drscmd
 import cmod.fmt as fmt
+import argparse
 import logging  # Additional settings required
-import copy
 
 if __name__ == '__main__':
   logging.root.setLevel(
@@ -62,60 +63,27 @@ if __name__ == '__main__':
   ])
   """
 
-  session = ss.GUISession([getset.set, getset.get, getset.wait, ss.shutdown])
-
-  # Duplicating the session to allow for default override.
-  prog_parser = copy.deepcopy(session.cmd.set.parser)
-
-  # Augmenting help messages
-  prog_parser.prog = 'gui_control.py'
-  prog_parser.add_argument('-h',
-                           '--help',
-                           action='store_true',
-                           help='print help message and exit')
-
-  ## Using map to store Default values:
-  default_overide = {
-      # Devices that can actually switch interfaces with the given string.
-      '--printerdev': '/dev/ttyUSB0',
-      '--camdev': '/dev/video0',
-      # The DRS4 and picoscope CANNOT actually be set by some device string, but
-      # requires a fake /dev/ path to be able to trigger the initialization
-      # routine
-      '--drsdevice': "/dev/MYDRS4",
-      '--picodevice': '/dev/MYPICOSCOPE',
-  }
-
-  for action in prog_parser._actions:
-    for option, default in default_overide.items():
-      if option in action.option_strings:
-        action.default = default
-
+  session = ss.GUISession([cmdbase.wait, ss.shutdown])
+  logger = logging.getLogger("SiPMCalibCMD.setup")
+  prog_parser = argparse.ArgumentParser("control.py",
+                                        "Starting an interactive CLI session",
+                                        add_help=True,
+                                        exit_on_error=True)
+  prog_parser.add_argument('--init',
+                           '-i',
+                           type=str,
+                           help="""File containing commands to run before the
+                                handing the interactive session to the users""")
   args = prog_parser.parse_args()
 
-  if args.help:
-    prog_parser.print_help()
-    sys.exit(0)
-
-  logger = logging.getLogger("SiPMCalibCMD.setup")
-  try:
-    logger.info("Running set command")
-    session.cmd.set.run(args)
-  except Exception as err:
-    logger.error(str(err))
-    logger.warning(
-        fmt.oneline_string("""
-          There was error in the device setup process, program will continue but
-          will most likely misbehave! Use at your own risk!"""))
-  try:
-    logger.info("Starting GPIO")
-    session.cmd.gpio.init()
-  except Exception as err:
-    logger.error(str(err))
-    logger.warning(
-        fmt.oneline_string("""
-          There was error in the GPIO setup, program will continue but will most
-          likely misbehave! Use at your own risk!"""))
+  # Running the initialization script
+  if args.init:
+    try:
+      session.cmd.onecmd(f'runfile {args.init}')
+    except Exception as err:
+      logger.error(
+          f"Error initializing using file {args.init}. We will continue, but the command session may misbehave!"
+      )
 
   session.start_session()  # Starting the session!
   # Notice that this will continue to run until the shutdown signal is sent from
